VCL-563 Added call to enable_dhcp in pre_capture.
Updated enable_dhcp. It was not working correctly if DHCP was enabled before it was called but the interface didn't have an IP address.


VCL-562
Added disable_user subroutine. Added get_user_names subroutine.

Added optional $user_password_only argument to set_password. This allows a password to be set without attempting to update passwords of any scheduled tasks or services which are configured to run as the user.


VCL-564
Updated run_script. It wasn't correctly retrieving the exit status of the script that was executed.

Added run_scripts subroutine. It searches directories under tools which match the OS and architecture of the image being loaded and then executes any scripts it finds.


VCL-561
Added check_image and update_imagerevision_info subroutines. These will eventually retrieve various inventory information from the image being loaded and store the information in the database. These aren't currently called.


Other
Update set_static_public_address to run the commands to set the address and configure DNS in a single command.