VCL-499
The move_virtual_disk subroutine would not work in the context of a vCenter host. That is,
the MoveVirtualDisk method is not implemented for these hosts. Instead, I followed the 
design used in the copy_virtual_disk subroutine, to test for a 'not implemented' string
in the $fault code. If so, execution proceeds to use the CloneVM method (via copy_virtual_disk 
and cleaning up source files afterwards).

The VIExt::http_get_file and VIExt::http_put_file methods sometimes fail on vCenter hosts, so
I wrapped those calls in a code_loop_timeout(...) subroutine. That way, repeated attempts will 
be made, up to a maximum of 50 elapsed seconds.