VCL-499
VMWare vCenter will modify the name of the virtual disk if the VM name is longer than
29 characters. This causes the actual location of the disk to be different than how
it is recorded in the VCL database. Now, when move/copy_virtual_disk is called,
the VM Name is shortened to an acceptable length while ensuing that it remains 
unique in the DB. The name is checked upon successful move/copy, and if it has 
changed, the database will be updated.