VCL-589 Updated Windows.pm::copy_capture_configuration_files to grant permissions to the SYSTEM account for the VCL directory in root's Cygwin profile. This allows scripts which run during Sysprep under the SYSTEM account to succeed.
Reworked Version_6.pm::run_sysprep. The Unattend.xml contents are now generated from the Perl code rather than copying a file from tools. This makes easier to troubleshoot.


VCL-503
Made changes to OS.pm::execute_new. Commands with long timeouts were failing. Removed timeout argument when Net::SSH::Expect object is created and added command to set the timeout value in the object every time the subroutine is called.

Added $identity_paths argument to call to execute_new in utils.pm::run_ssh_command.


VCL-573
Removed the retrieval of user info for users specified in imagemeta.usergroupid from get_request_info and get_imagemeta_info. Removed entries from DataStructure.pm which return this information.


VCL-545
Changed path where virtual disk files are located in the datastore. Instead of each residing in its own directory, files are stored at the root of the datastore. This was done because virt-manager assumes they reside in the root directory and cannot traverse subdirectories.

Commented out code in KVM.pm::copy_virtual_disk which attempts to first convert the image to raw format. This was not working consistently.


Other
Updated Windows.pm::delete_file to detect if the path contains a wildcard. Added an additional command to delete files/directories beginning with a period.

Added condition to utils.pm::is_request_deleted to check if the request state is makeproduction. This is effectively the same as deleted, but the user chose to update the production revision of the image when the request was ended. Also added a check to vcld where it checks if a request is deleted.

Commented out line which returns cached user information in utils.pm::get_user_info. User preference changes were not being applied because the cached information was returned.