VCL-576 Finalizing for 2.3 release
added $VCLversion global variable that is set in index.php as the current version of VCL. This is then used when javascript files are included in the page headers as an appended query variable in the form of somefile.js?v=$v where $v is locally set to $VCLversion. This makes browsers reload the javascript file whenever $VCLversion is changed so that people don't end up with old, cached javascript files with new code, resulting in things not working correctly. For whatever, we're regularly seeing apache and browsers think the javascript files have not been modified since the browser last loaded them when they actually have.