VCL-450 Set $ENV{PERL_LWP_SSL_VERIFY_HOSTNAME} = 0 in vSphere_SDK.pm::initialize. The vSphere SDK is might be unable to connect some hosts due to SSL signing problems.
Uncommented lines to set sourceDatacenter and destDatacenter in vSphere_SDK.pm::copy_virtual_disk. With vCenter 5, the CopyVirtualDisk function was failing and generating an "Invalid Argument" fault instead of a "not implemented" fault. As a result, the subroutine was returning before it attempted the clone method. Uncommenting the lines caused a "not implemented" fault which is what is expected.

Updated vSphere_SDK.pm::_get_datastore_info. Under vCenter 5 with DRS enabled, the datastore URLs begin with ds://. The subroutine wasn't checking for this.  As a result, the ds: path wound up in the vmx and snapshots failed.

Added additional argument to call to get_image_info in vSphere_SDK.pm::_clean_vm_name to not display warnings if the image does not exist.