VCL-545 Fixed problems with base image capture under KVM. Updated image.pm to correctly determine the installtype.
Added code to determine the running domain when capturing an image rather than relying on the domain name matching the computer name.

Changed capture code to no longer save a copy of the XML file which defines the domain in the datastore and repository directories. This can cause problems for the code used to find images. Instead, a copy is saved in the image. This isn't currently used but will be useful in the future to customize hardware based on what was actually in the captured image.


VCL-450
Added code to VIM_SSH.pm to check if the VM is already powered off before attempting to power it off.


Other
Updated Linux.pm::create_directory to detect if ~ was passed as the argument. It skips directory creation if this is the case. Otherwise, it was creating a '~' directory under the home directory.

Added $ignore_error argument to utils.pm.