VCL-545 Updated libvirt.pm. Added missing calls to insertloadlog. Fixed a few bugs with image capture. The code renames the image so that images are not inappropriately named vmware*. It was setting this updated name in the database but incorrectly naming image files saved to the datastore and repository. Updated code to name the files correctly.
Other
Updated image.pm so that a warning is not displayed if the sysadmin email address is not configured.

Added code to Linux.pm::pre_capture to allow the provisioning module to specify the "end_state" -- meaning if a hash reference is passed with an 'end_state' key set to on, the computer is not powered off by the OS module. This is already in Windows.pm but was missing from Linux.pm.