VCL-623 Added Windows.pm::get_public_interface_name. This calls OS.pm::get_public_interface_name and adds a loop if the interface name can't be determined. This should solve the problem where the code attempts to determine the public interface name before the interface driver is installed and initialized.
Changed OS.pm::get_public_interface_name so that it never returns the private interface name if that interface is only assigned the private IP address. It may return the private interface name if that interface is assigned other addresses. This is done for future enhancements when a single interface is supported.

Commented out section in get_network_configuration which was looping in order to attempt to determine if devices were being installed. This functionality should be replaced by the new code in this commit.

Updated Module.pm::code_loop_timeout to return the value returned by the code reference instead of a simple boolean value. Also added an additional argument to be specified controlling how often log messages are displayed.