VCL-624
The getUserlistID function takes a user's loginid as its first parameter. Because of the way findAffilFunc is written, it is possible to hand the function a unityid and it doesn't die, but it is still incorrect. That is, it doesn't die as long as the
system uses the testGeneralAffiliation function in the findAffilFuncs array.