VCL-560
For installations using the vmprofile.password value in the database, this value is stored in cleartext. That means that if there are multiple management nodes, each
management node will have access to this password, which may not be desirable. The password is also passed in clear text along the network between the mysql database and the
management node(s). 

This implements the ability to set an asymmetric key so that the password can be set in the website but only be decypherable from the management node(s) with the private key.