VCL-678 Removed and/or commented out several unnecessary notify messages in Module.pm to reduce vcld.log output.
VCL-707
Changed how vcld.log lines are deliminated. All fields are now separated by a pipe character. If a field is blank, a pipe character is added. This allows output to be pasted into a spreadsheet and text to columns results in the fields in the correct column for all lines.

VCL-16
Moved initial user connection check from reserved.pm to inuse.pm. This removed duplicate code.

Updated reserved.pm so that the parent reservation waits for all child reservations' reserved processes to terminate before changing the request state to inuse. This fixes a problem where the state would change before the child's reserved process ended. The child's inuse process would fail to launch because vcld detected the other process still running.

Moved user connection check loop from OS.pm::is_user_connected to inuse.pm. Updated inuse.pm to update reservation.lastcheck for all reservations in a cluster when a user connection is detected. While looping waiting for a connection, the current reservation.lastcheck is retrieved from the database. If it changed it means some other reservation in the cluster detected a connection and the loop terminates. 

Reworked and simplified much of inuse.pm::process. It was still using goto calls.