VCL-582 Updated SysV.pm to not display a warning when trying to stop a service which is already stopped.
Updated add_ext_sshd_service to make a backup of the original /etc/init.d/sshd file. The sshd file is edited to change the 'status $SSHD' line to 'status -p $PID_FILE openssh-daemon'. This allows 'service sshd status' to report the correct status if sshd is stopped and ext_sshd is running.