VCL-725 Reworked utils.pm::run_command to call the command via backticks instead of opening a pipe. This seems to always return the correct exit status.
VCL-726
Added call to vcld:: make_new_child which sets the process group of vcld state processes to that of the forked state process. Without doing this, the process group of the forked process is that of the parent vcld daemon process.

Added kill call to REAPER which kills all processes in the process group. Since the process group was set for the forked vcld state process, all processes it forks will be in its process group. This causes forked processes to be killed when a vcld state process is killed.