VCL-698
skip useradd if user's home directory is on network mounted storage. Prevousily this steps are to delete and then try to re-add the user's account even if it was network mounted.
this might need to be re-addressed in the future if new accounts need to be created on network related storage. The basic assumption is that the network based home dir would already exist.