VCL-734 Updated new.pm::computer_not_being_used to not use recursion. This could cause recursive loops. Added a multiple attempt loop instead.
Moved command to rename the forked process from State.pm::initialize to vcld::make_new_child. This was done to have the rename occur as early as possible. Other reservations run pgrep and use the process name to determine if another process is running for a reservation. This allows a running process to be detected sooner after a process begins.

Added Semaphore.pm::get_process_semaphore_ids. This is used in new.pm::computer_not_being_used to detect if a running reload process may be transferring an image. It's possible for old, stale semaphore lockfiles to exist in /tmp. This subroutine verifies that the lockfile matches the PID of the other running process.

Added check to State.pm::initialize if it fails to update the state to pending. There are a few cases where this is normal such as if the request was deleted in the brief time before it could be updated to pending.

Updated utils.pm::reservation_being_processed to return either an array or integer.