VCL-767 Added update_computer_private_ip_address subroutine to utils.pm. Renamed update_computer_address to update_computer_public_ip_address and updated calling subroutines in OS.pm.
VCL-753
Reworked enable_firewall_port and disable_firewall_port in Linux.pm. The clean_iptables subroutine was not working correctly because the old code didn't allow "any" to be passed for the scope.

Other
Alphabetized export list in utils.pm. Please put subroutine names in the correct location when adding to this and reorder it if you rename a subroutine.