VCL-584 Added functionality to backend to allow an image to be captured and then the reservation continued.
Added check to determine if request state is 'checkpoint' in image.pm::process. If the state is checkpoint, the computer isn't reloaded at the end of the capture. Instead, the computer is powered on and OS module's post_load and reserve subroutines are called. The request state is then changed to reserved.

Added utils.pm::update_request_checkuser. This is called from image.pm for checkpoint captures to prevent the reservation from timing out after the capture os complete in case anything went wrong.

Updated capture subroutines in libvirt.pm, VMware.pm, and vbox.pm to not delete the VM after an image is captured if the state is checkpoint.


VCL-767
Updated utils.pm::update_computer_private_ip_address to allow either a computer ID or hostname argument to be specified.

Added utils.pm::get_computer_private_ip_address_info. It is used to retrieve all of the private IP addresses from the database.

Added utils.pm::hostname_to_ip_address. It calls gethostbyname. It will be used to check if the private IP addresses in the database match what the hostnames resolve to. It is also used to determine the management node's private IP address instead of reading /etc/hosts.


Other
Added argument to get_user_sshPublicKeys in OS.pm so a warning isn't displayed if the value is not set for a user.