VCL-774
When using vSphere provisioning, the _get_file_info subroutine is unnecessarily slow (measured in minutes rather than seconds) especially on systems with many VMs (e.g. vCenter clusters). This is due to an overly comprehensive query definition, when in fact only VmDiskFileQuery, FileQuery and FolderFileQuery need to be used.
This commit eliminates the unnecessary query parameters.