VCL-685
Updated VIM_SSH.pm::_services_restart and _check_service_pid. Occasionally '1' was returned as the PID of a service and the code had been writing this to the .pid file. This caused problems. Added a check for a valid value. Added additional services to check. 

Updated VMware.pm::is_vm_registered to accept datastore formatted paths.

Updated VMware.pm::delete_vm to remove the VM's parent directory if it was named after the VM.

Added VIM_SSH.pm::vm_suspend for a script I wrote to migrate VMs. This subroutine may be used at some point in the future.

Updated VMware.pm::get_datastore_info to strip away the leading 'ds://' in the vmdk's URL if it was present. This allows paths containing the URL to be used.

Updated vSphere_SDK.pm::copy_virtual_disk to delete the source VM it creates during a cloning operation if the clone fails.

VCL-724
Added VMware.pm::check_multiextent and vSphere_SDK.pm::is_multiextent_disabled. These are called by the copy/move_vmdk subroutines to detect if a 2gbsparse vmdk operation fails because multiextent is not enabled on the host.