VCL-776 - rework resource code to have a base class for all resources and inheriting classes for each resource type
image.js: modified saveResource: added code to expand Advanced Options section if there is an error with one of the form elements inside it and then to put the focus on that item so that the error popup gets displayed. had a problem where submitting changed to an image with Advanced Options closed but with an error in that data made it seem like the Save Changes button did not work.