VCL-780 - combine new reservation and current reservations pages
fixed a problem where when the Reservations page is initially loaded, an image only has 1 revision. So, the data for imagestore contains 'revisions' = 0. However, if a user creates a new revision and never leaves the page, once the revision has been captured, the user will not be prompted to select a revision. Fixed this by setting 'revisions' in imagestore for the image to 1. There is a problem that needs to be fixed later where the new revision will not show up in the prompting until it is finished and has 'deleted' set to 0. This results in a user being prompted to select a revision when only 1 revision exists until the catpure is finished.

image.php: modified AJupdateImage: send imageid back with returned data

requests.php: modified newReservationHTML: changed imagestore from ItemFileReadStore to ItemFileWriteStore so we can change the value of 'revisions'

image.js: modified submitUpdateImageClickthroughCB: update value of 'revisions' for entry in imagestore to = 1 so that user will get prompted to select an image revision