VCL-174 Changed iptables.pm::initialize to only check for the iptables command, not the service. The command can be used if firewalld is used.
Updated configure_nat to figure out the internal network address and network bits. The MASQUERADE rule was tightened so that it matches only if the destination address is not on the internal network to allow things to work if the public and internal networks are using the same interface.

Added check to Windows.pm::grant_access when process_connect_methods is called. It was not returning if process_connect_methods failed. As a result, the reservation did not fail when it should have.

Other
Updated Linux.pm::update_public_hostname. It was generating unnecessary warnings if the public IP address did not resolve. Added utils.pm::ip_address_to_hostname. This is called by update_public_hostname instead of running ipcalc. If the hostname can't be resolved the hostname is set to the name in the VCL database.

Updated Linux.pm::set_static_public_address to continue to try to set the default route even if the IP address is correct to begin with.