VCL-392 Added accessor to imagemeta.sethostname to DataStructure.pm.
Added Windows.pm::set_computer_hostname subroutine. Updated Windows.pm::post_load to set the computer hostname if imagemeta.sethostname = 1. If the hostname is changed, the computer is rebooted during post_load. If a reboot is to be done, a scheduled task to execute update_cygwin.cmd is added to run at startup. This task is also deleted during pre_capture to avoid conflicts with other scripts.

Added condition to Linux.pm::post_load to call update_public_hostname only if imagemeta.sethostname is not set to 0. The subroutine is called if imagemeta.sethostname is undefined or set to 1.