VCL-174 - NAT - support for sites that have small IP address ranges
update-vcl.sql: removed insert that populated nathost table with existing management nodes. This was initially added before there was a way to add nat hosts and all added management nodes would also be added as nat hosts. The code was added to be able to set any management node or computer as a nat host. So, this initial population of the nathost table with management nodes is unnecessary.