VCL-834 - Disable timeout for disconnected users is meaningless for imaging reservations
I listed the wrong JIRA issue in the last commit. I added a newline at the bottom of the file to do another commit.

requests.js: modified selectResType: only show "Disable timeout for disconnected users" for basic reservations; previously it was shown for basic and imaging reservations; imaging reservations are not timed out when users are disconnected. So, having this option for them didn't make sense.