VCL-845 Removed reference to blockRequest.admingroupid from utils.pm::get_management_node_blockrequests and from DataStructure.pm.
VCL-847
Fixed problem in utils.pm::reservation_being_processed. The $parent_reservation_id variable was getting initialized every time. Later on it is set to the actual ID if not defined. As a result, it would never be set to the correct reservation ID.

Added State.pm::wait_for_reservation_loadstate. Updated reserved.pm so that child reservations wait for the parent to generate an "acknowledgetimeout" computerloadlog entry. The timestamp of this entry is used to determine when to stop checking for acknowledgment. 

Updated reclaim.pm to call state_exit instead of switch_state.

Commented out section in State.pm::wait_for_child_reservations_to_exit where it exited immediately if the request was deleted. This causes problems with cluster requests. Added a comment explaining the problem for future reference.

Updated how State.pm::state_exit handles setting request.state if the current state is deleted. It was skipping the request.state update instead of changing it to completed.

Other
Added code to new.pm::wait_for_child_reservations to use a short interval between checks at the start, and then increase it. This speeds things up when all child reservations are ready early on.