VCL-844 Added subroutines to VMware.pm: * get_vmx_url_base_directory_path * get_vmx_intermediate_directory_path * _get_datastore_url * _get_url_path * configure_vmhost_ssh_keys * copy_vmhost_ssh_public_key_to_another_host * copy_file_to_another_host * setup_migrate_vm * migrate_vm * migrate_revert_source
Added subroutines to utils.pm:
* update_computer_vmhost_id.
* setup_get_hash_multiple_choice.

Updated Module.pm::new to call initialize on ManagementNode.pm objects even if $SETUP_MODE is true.

Updated Module.pm::create_object to accept an additional argument which allows arguments to be passed to the constructor of the object being created as well as the DataStructure.pm object.

Updated Module.pm::create_vmhost_os_object to accept an argument specifying which VM host to create an object for.

Updated DataStructure.pm to accept a vmhost_identifier argument rather than use a vmhost_id argument.

Changed where cached data is stored from %ENV to $self in order to allow it to work with multiple VMware.pm objects.


VCL-851
Added check to skip the attempt to use vSphere_SDK.pm if vmprofile.password is not configured.