VCL-844 Added subroutines to VIM_SSH.pm: firewall_ruleset_allow_ip firewall_ruleset_enable get_firewall_ruleset_info get_matching_firewall_ruleset_info is_firewall_port_allowed
These are used to determine configure the firewalls between 2 ESXi hosts to allow SSH traffic for a migration.

Fixed problem where a VMware snapshot would fail if taken on a powered off or suspended VM after a previous snapshot was successfully taken.

Added 6 minute timeout to vim-cmd power on command. Powering on a VM may take some time if it is suspended.