VCL-967 Added subroutines to libvirt.pm: get_node_network_info get_node_network_xml_string get_node_interface_info get_node_interface_xml_string
Added check to libvirt.pm::initialize to make sure the VM networks configured in the VM profile actually exist on the VM host. If not, a critical message is sent.

Updated libvirt.pm::generate_domain_xml to check if the VM network matches a virtual network on the host. If so, the interface type gets set to 'network'. Otherwise, it gets set to 'bridge'.