VCL-867 Made another attempt to update DataStructure.pm::_automethod to not display warnings when get_image_domain* is called and the data is not defined. The previous change wasn't catching all calls.
Added utils.pm::get_active_directory_domain_credentials. This is needed to determine the domain username and password in order to remove a computer from a domain after the preloaded image was reconfigured to not join a domain. The normal $self->data->get_image_domain* information is not available in this case.

Updated Provisioning.pm to check if OS module implements a node_status_os_check subroutine and call it if it does.

Added Windows.pm::ad_check. It contains all logic to determine if an image is configured for AD, if a computer is already joined to a domain, and calls the necessary subroutines to either unjoin the domain, join the domain, or both if the computer is joined to a different domain or located in the wrong OU. Removed similar logic from ad_join.

Replaced call from ad_join to ad_check in Windows.pm::post_load.

Added Windows.pm::node_status_os_check which simply returns the value of ad_check. Naming reason: ad_check is called elsewhere and Windows-specific. node_status_os_check is more general and may perform additional functions in the future.

Reworked Windows.pm::ad_unjoin to call wmic.exe rather than building Powershell script. Need to test this and determine which is better.

Updated Windows.pm::ad_search, ad_delete_computer, and ad_search_computer to accept domain DNS name argument. This is used when a computer is joined to a domain but shouldn't be (image was reconfigured after computer was loaded).