VCL-1033 Improved Windows.pm::post_load to attempt to retrieve or set the public IP before settings a persistent default route.
Added code in post_load if the retrieval failed and DHCP is used to call 'ipconfig /renew' and check again.

Cleaned up and modernized Windows.pm::ipconfig_renew.

Changed Version_6.pm::wait_for_response to check if the computer is a VM. If so, the total wait time is reduced to 5 minutes from 10-30. If the wait time is reached without a response, an attempt is made to call the provisioning object's power_reset subroutine, then another response loop is attempted.