VCL-971 Updated ufw.pm::save_configuration to add the '*<table>' lines as it finds them which could be intermingled with lines containing 'vcl'. It was previously assuming everything belonged in 'filter'.
Updated iptables.pm::get_table_info to handle text in the jump section of a rule that contains special characters such as square brackets. The code was previously passing what it found to a s/// regex. This was failing if the text contained special regex characters. Rewrote regex which locates the jump section to exclude any part of a subsequent match section.

Tested using Ubuntu/ufw for both a NAT host and client and it worked fine.