VCL-1109 - date duplicated on statistic graph labels for day after daylight saving time rolls back
The fix mentioned in VCL-1109 of just adding a value to the label doesn't really fix the underlying problem and give incorrect values on the graphs. A better fix was to change the timezone to UTC before doing any date work so that DST isn't used.

statistics.php: modified getStatGraphDayData, getStatGraphDayConUsersData, getStatGraphConBladeUserData, and getStatGraphConVMUserData: added a call at the beginning of each function to get the current timezone, then set the timezone to UTC; added a call before any returns in each function that sets the timezone back to was was retreived at the beginning of the function