VCL-1125 - update libvirt provisioning module to support image project
libvirt.pm: modified generate_domain_xml: added code to check for image project not being "vcl", and if it is not, check for other networks on host to connect eth1 to. Also check for additional NICs needing to be created based on intersection of image project value and network names.